refactor(api): extract shared serialize_fido2_options utility
